# src/main/resources/swagger/api.yml
...
openapi: 3.0.1
info:
  title: Public Bon API
  description: >
    The bonlimousin.com public API for retrieving livestock information and content
  contact:
    name: Johan Frostmark
    url: http://www.bonlimousin.com
    email: johan@bonlimousin.com
  license:
    name: unknown
  version: 0.0.1
servers:
  - url: http://localhost:9000
    description: Development server
  - url: https://localhost:9000
    description: Development server with TLS Profile
...
paths:
  /api/public/cows:
    get:
      tags:
        - cowvo-resource
      summary: Find all cows matching criteria
      description: ''
      operationId: findCowVOs
      parameters:
        - name: page
          in: query
          description: Page number of the requested page
          schema:
            type: integer
            format: int32
            minimum: 0
            default: 0
        - name: size
          in: query
          description: Size of a page
          schema:
            type: integer
            format: int32
            minimum: 1
            default: 10
        - name: sort
          in: query
          description: 'Sorting criteria in the format: property(,asc|desc). Default
            sort order is ascending. Multiple sort criteria are supported.'
          style: form
          explode: true
          schema:
            type: array
            items:
              type: string
        - name: linageId.equals
          in: query
          schema:
            type: integer
            format: int32
        - name: birthDate.greaterThan
          in: query
          schema:
            type: string
            format: date-time
        - name: birthDate.lessThan
          in: query
          schema:
            type: string
            format: date-time
        - name: gender.equals
          in: query
          schema:
            type: string
            enum:
              - HEIFER
              - BULL
        - name: hornStatus.in
          in: query
          style: form
          explode: true
          schema:
            type: array
            items:
              type: string
              enum:
                - UNKNOWN
                - HORNED
                - POLLED
                - SCURS
                - DEHORNED
                - DISBUDDED
        - name: matriId.equals
          in: query
          schema:
            type: integer
            format: int32
        - name: patriId.equals
          in: query
          schema:
            type: integer
            format: int32
        - name: weight0.greaterThan
          in: query
          schema:
            type: integer
            format: int32
        - name: weight0.lessThan
          in: query
          schema:
            type: integer
            format: int32
        - name: weight0.specified
          in: query
          schema:
            type: boolean
        - name: weight200.greaterThan
          in: query
          schema:
            type: integer
            format: int32
        - name: weight200.lessThan
          in: query
          schema:
            type: integer
            format: int32
        - name: weight200.specified
          in: query
          schema:
            type: boolean
        - name: weight365.greaterThan
          in: query
          schema:
            type: integer
            format: int32
        - name: weight365.lessThan
          in: query
          schema:
            type: integer
            format: int32
        - name: weight365.specified
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CowVO'
        '404':
          description: Not Found
          content: {}
        '501':
          description: Not Implemented
          content: {}
      deprecated: false
  '/api/public/cows/{earTagId}':
    get:
      tags:
        - cowvo-resource
      summary: getCowVO
      description: ''
      operationId: getCowVO
      parameters:
        - name: earTagId
          in: path
          description: Numeric identifier for cattle
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CowVO'
        '404':
          description: Not Found
          content: {}
        '501':
          description: Not Implemented
          content: {}
      deprecated: false
...
  '/api/public/cows/{earTagId}/photographs':
    get:
      tags:
        - cowvo-resource
      summary: Get all PhotographVOs for a cow
      description: ''
      operationId: getAllPhotographVOsByCow
      parameters:
        - name: earTagId
          in: path
          description: Numeric identifier for cattle
          required: true
          schema:
            type: integer
            format: int64
        - name: page
          in: query
          description: Page number of the requested page
          schema:
            type: integer
            format: int32
            minimum: 0
            default: 0
        - name: size
          in: query
          description: Size of a page
          schema:
            type: integer
            format: int32
            minimum: 1
            default: 10
        - name: sort
          in: query
          description: 'Sorting criteria in the format: property(,asc|desc). Default
            sort order is ascending. Multiple sort criteria are supported.'
          style: form
          explode: true
          schema:
            type: array
            items:
              type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PhotographVO'
        '404':
          description: Not Found
          content: {}
        '501':
          description: Not Implemented
          content: {}
      deprecated: false
  '/api/public/cows/{earTagId}/pictures':
    get:
      tags:
        - cowvo-resource
      summary: Get all PictureVOs for a cow
      description: ''
      operationId: getAllPictureVOsByCow
      parameters:
        - name: earTagId
          in: path
          description: Numeric identifier for cattle
          required: true
          schema:
            type: integer
            format: int64
        - name: page
          in: query
          description: Page number of the requested page
          schema:
            type: integer
            format: int32
            minimum: 0
            default: 0
        - name: size
          in: query
          description: Size of a page
          schema:
            type: integer
            format: int32
            minimum: 1
            default: 10
        - name: sort
          in: query
          description: 'Sorting criteria in the format: property(,asc|desc). Default
            sort order is ascending. Multiple sort criteria are supported.'
          style: form
          explode: true
          schema:
            type: array
            items:
              type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PictureVO'
        '404':
          description: Not Found
          content: {}
        '501':
          description: Not Implemented
          content: {}
      deprecated: false
  '/api/public/cows/{earTagId}/pictures/{pictureId}/{name}':
    get:
      tags:
        - cowvo-resource
      summary: Get image for a cow
      description: ''
      operationId: getImageForCow
      parameters:
        - name: earTagId
          in: path
          description: Numeric identifier for cattle
          required: true
          schema:
            type: integer
            format: int64
        - name: pictureId
          in: path
          description: Numeric identifier for picture
          required: true
          schema:
            type: integer
            format: int64
        - name: name
          in: path
          description: Name of photo
          required: true
          schema:
            type: string        
      responses:
        '200':
          description: OK
          content:
            'image/*':
              schema:
                type: string
                format: binary
        '404':
          description: Not Found
          content: {}
        '501':
          description: Not Implemented
          content: {}
      deprecated: false
...
components:
...
  securitySchemes:
    jwt:
      type: http
      description: JWT Authentication
      scheme: bearer
      bearerFormat: JWT
  schemas:
    CowVO:
      title: CowVO
      type: object
      properties:
        earTagId:
          type: integer
          format: int32
        linageId:
          type: integer
          format: int64
        linageName:
          type: string
        name:
          type: string
        birthDate:
          type: string
          format: date-time
        gender:
          type: string
          enum:
            - HEIFER
            - BULL
        hornStatus:
          type: string
          enum:
            - UNKNOWN
            - HORNED
            - POLLED
            - SCURS
            - DEHORNED
            - DISBUDDED
        matriId:
          type: integer
          format: int32
        patriId:
          type: integer
          format: int32
        weight0:
          type: integer
          format: int32
        weight200:
          type: integer
          format: int32
        weight365:
          type: integer
          format: int32
        storyHandle:
          type: string
        visibility:
          type: string
          enum:
            - ROLE_ADMIN
            - ROLE_USER
            - ROLE_ANONYMOUS
...
    PhotographVO:
      title: PhotographVO
      type: object
      properties:
        caption:
          type: string
        earTagId:
          type: integer
          format: int32
        height:
          type: integer
          format: int32
        id:
          type: integer
          format: int64
        image:
          type: string
          format: byte
        imageContentType:
          type: string
        taken:
          type: string
          format: date-time
        width:
          type: integer
          format: int32
        visibility:
          type: string
          enum:
            - ROLE_ADMIN
            - ROLE_USER
            - ROLE_ANONYMOUS
    PictureVO:
      title: PictureVO
      type: object
      properties:        
        id:
          type: integer
          format: int64                                       
        caption:
          type: string
        taken:
          type: string
          format: date-time        
        visibility:
          type: string
          enum:
            - ROLE_ADMIN
            - ROLE_USER
            - ROLE_ANONYMOUS
        sources:
          type: array
          items:
            $ref: '#/components/schemas/PictureSourceVO'
    PictureSourceVO:
      title: PictureSourceVO
      type: object
      properties:        
        name:
          type: string
        url:
          type: string            
        contentType:
          type: string
        width:
          type: integer
          format: int32
        height:
          type: integer
          format: int32                
...
security:
  - jwt: []